style: naming conventions, bug fixes, and code cleanup (#23)
* feat: swapchain, command buffers, frame sync, and clear screen

Implement Etape 3 — triple-buffered swapchain with MAILBOX present
mode, double-buffered frame synchronisation (fences + semaphores),
dynamic rendering clear to dark teal, and Synchronization2 barriers.

- Add gpu::Swapchain class (B8G8R8A8_SRGB, old-swapchain reuse)
- Add physical_device() getter to gpu::Device
- Use Signal<ResizeEvent> for window resize communication
- Add NOMINMAX and WIN32_LEAN_AND_MEAN globally in CMake
- Remove redundant local Windows.h guards from gpu_surface.cpp


* fix: enable synchronization2 feature at device creation

The Vulkan validation layer flagged vkCmdPipelineBarrier2 and
vkQueueSubmit2 calls because the synchronization2 feature was
not enabled. Chain VkPhysicalDeviceSynchronization2Features into
the device creation pNext chain alongside dynamic rendering.


* removed Github copilot from extensions

* fix: bug fixes, naming conventions, and STYLE.md enforcement

- Fix dangling pointer: family_indices scoped to outlive swapchain creation
- Fix orphaned semaphore: suboptimal acquire now proceeds to present
- Fix 0x0 swapchain recreation on minimise (3 call sites guarded)
- Fix image barriers: use VK_QUEUE_FAMILY_IGNORED instead of 0
- Fix missing Vulkan 1.3 physical device check in rateDevice()
- Fix spurious WM_SIZE during Win32 window construction
- Fix PeekMessageW filtering (was dropping WM_QUIT)
- Fix XCB centring underflow and WM_NORMAL_HINTS type atom
- Fix WindowEvent union zero-initialisation (latent UB)
- Replace all throw statements with std::cerr + std::abort()
- Enforce camelCase for functions/methods across entire codebase
- Enforce m_ prefix for member variables, drop trailing underscore
- Rename namespaces: Gpu, WindowLib, SignalLib, TestFixtureLib
- Update STYLE.md with exception policy, naming conventions


* style: review fixes, Signal<T> encapsulation, licence headers

- Add return after std::abort() in window library (4 locations)
- Make Signal<T> members private (struct → class)
- Add [[nodiscard]] to Window accessors and factory
- Remove unused #include <stdexcept> from gpu_surface.cpp
- Update licence headers to 4-space indented block comment style


* style: runAll returns bool, doxygen block comments, auto removal

- Change runAll() return type from int to bool
- Use static_cast<int> at call sites in all 3 test mains
- Replace auto with explicit types in test_fixture.cpp
- Use /*! */ block style for Signal<T> class doxygen
- Convert // comments to //! doxygen in window.hpp
- Minor spacing and blank line cleanup


* style: break before binary operators in clang-format

Change BreakBeforeBinaryOperators from None to All so that
wrapped expressions indent with continuation width rather
than aligning far right on operators like &&, ||, +.


* style: tighten include groups and merge doxygen lines

- Remove blank lines between include groups in src/ headers and sources
- Merge multi-line //! doxygen comments into single lines
- Binary operator line break now before | (clang-format All)


* refactor: remove Gpu namespace and drop gpu_ file prefix

The Gpu namespace adds unnecessary qualification at this stage since
all source files are in src/ with no risk of name collision. Simplifies
call sites (Instance, Device, Swapchain instead of Gpu::Instance, etc.)
and file names (instance.hpp instead of gpu_instance.hpp).

- Remove namespace Gpu from all 8 source files, un-indent content
- Rename gpu_instance → instance, gpu_device → device,
  gpu_surface → surface, gpu_swapchain → swapchain
- Update all #include directives and CMakeLists.txt
- Convert remaining multi-line //! to /*! */ block doxygen


* fix: correct STYLE.md member convention and runAll() doxygen

- STYLE.md: member variables are m_snake_case (no trailing underscore)
- test_fixture.hpp: update runAll() comment to match bool return type


* refactor: move Window types into WindowLib namespace

Move WindowEvent, WindowConfig, Window, Win32Window, and XcbWindow
into the WindowLib namespace for consistency with SignalLib and
TestFixtureLib. Update all call sites with WindowLib:: qualifiers.


---------

+++ libs/signal/tests/signal_tests.cpp
@@ -96,5 +96,5 @@ TEST_CASE(thread_safety)
 
 int main()
 {
-    return TestFixtureLib::runAll();
+    return static_cast<int>(TestFixtureLib::runAll());
 }
